fix: valor do cartão de crédito agora pode ser decimal

+++ src/main/webapp/paginas/cliente/compra/js/secaoCartaoCredito/SelecaoCartaoCredito.js
@@ -46,7 +46,7 @@ export class SelecaoCartaoCredito extends HTMLDivElement {
             if (c.id == id){
                 cartao = { 
                     idCartaoCredito: id,
-                    valor: parseInt(
+                    valor: parseFloat(
                         this.querySelector('input').value, 10
                     ),
                 };
